pyproject.toml
[tool.poetry]
name = "gallery"
version = "1.0.0"
description = ""
authors = ["Nikita Muzykantov <mzkntv@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
